Remove the global points-per-second scale
Nothing chooses one scale for a whole series any more — the window does it, per
window. The median and percentile measurements survive because the caption and
the zoom limit still need them.

WaterfallChartStyle.bar(...) and .drawnEnd(...) were also fully dead — a
previous task had already removed both surfaces' only call sites and left the
implementations behind. Removed them along with the six tests that pinned
drawnEnd's floor; the one that also checked valueLabel's independence from a
drawn width was already covered by TrafficStatsViewModelTests.

WaterfallTimeScale is renamed to WaterfallDurations: with the scale gone, only
two pure statistics functions and the tailPercentile constant survive, and
every real call site (WaterfallViewModel and its tests, plus this file's own
tests) moves in this commit.

WaterfallViewModel.scale(visibleWidth:) is gone too — it only ever built the
now-deleted WaterfallTimeScale.make(...) and had no caller left once that was
removed.
