+++ Tests/ScytherTests/Features/WaterfallChartStyleTests.swift
@@ -56,71 +56,6 @@ final class WaterfallChartStyleTests: XCTestCase {
                              "a zero-wide axis has nothing to draw on")
     }
 
-    // MARK: - Minimum rendered width
-
-    /// A request too short to draw is given exactly one point of ink — not a fraction of the
-    /// axis, which is what it was. That version's floor grew with the session: over five minutes
-    /// it inflated every bar to more than three seconds, so a 5 ms request and a 3 s request drew
-    /// identically and a floored bar could reach across a request it never ran alongside.
-    func testABarTooNarrowToSeeIsDrawnExactlyOnePointWide() {
-        let bar = entry(start: 4, duration: 0.0002)
-        let secondsPerPoint = 60.0 / 180.0
-        let width = WaterfallChartStyle.drawnEnd(of: bar, upperBound: 60, plotWidth: 180) - bar.start
-        XCTAssertEqual(width / secondsPerPoint, Double(WaterfallChartStyle.minimumBarWidth),
-                       accuracy: 0.0001)
-    }
-
-    /// The guard the axis-fraction version could not offer: however long the session runs, the
-    /// floor adds at most one point of ink, so it cannot invent an overlap a reader could
-    /// otherwise have ruled out.
-    func testTheFloorNeverAddsMoreThanOnePointOfInk() {
-        let secondsPerPoint = 405.0 / 178.0
-        for duration in [0.0, 0.0005, 0.005, 0.5, 5.0, 50.0] {
-            let bar = entry(start: 10, duration: duration)
-            let drawn = WaterfallChartStyle.drawnEnd(of: bar, upperBound: 405, plotWidth: 178)
-            let addedPoints = (drawn - (bar.start + duration)) / secondsPerPoint
-            XCTAssertLessThanOrEqual(addedPoints, Double(WaterfallChartStyle.minimumBarWidth) + 0.0001,
-                                     "a \(duration)s bar was inflated by \(addedPoints) points")
-        }
-    }
-
-    /// The teeth against sliding back to a fraction of the axis: a floor expressed in points
-    /// shrinks in seconds as the plot gets wider, and a floor expressed as a fraction does not.
-    func testAWiderPlotMakesTheFloorWorthLessTime() {
-        let bar = entry(start: 0, duration: 0)
-        let narrow = WaterfallChartStyle.drawnEnd(of: bar, upperBound: 100, plotWidth: 100)
-        let wide = WaterfallChartStyle.drawnEnd(of: bar, upperBound: 100, plotWidth: 400)
-        XCTAssertEqual(narrow / wide, 4, accuracy: 0.0001,
-                       "four times the plot, a quarter of the seconds")
-    }
-
-    /// The floor only ever grows a bar that could not be seen. A bar with real length is drawn at
-    /// exactly the length it ran, or the chart stops being a measurement.
-    func testABarWithRealLengthIsDrawnAtItsTrueLength() {
-        let bar = entry(start: 1, duration: 5)
-        XCTAssertEqual(WaterfallChartStyle.drawnEnd(of: bar, upperBound: 10, plotWidth: 200), 6,
-                       accuracy: 0.0001)
-    }
-
-    /// The preview's contract. Charts sizes its leading axis to its own labels, so that surface
-    /// cannot state a plot width; it passes zero and gets the true lengths it shipped with.
-    func testASurfaceThatDoesNotKnowItsPlotWidthDrawsTrueLengths() {
-        let bar = entry(start: 2, duration: 0.0001)
-        XCTAssertEqual(WaterfallChartStyle.drawnEnd(of: bar, upperBound: 405, plotWidth: 0),
-                       2.0001, accuracy: 0.000001)
-    }
-
-    /// Widening the bar must not touch what it says. Under a floor worth three seconds, a two
-    /// millisecond request still reports two milliseconds.
-    func testAFlooredBarStillReportsItsRealDuration() {
-        let bar = entry(start: 0, duration: 0.002)
-        let drawn = WaterfallChartStyle.drawnEnd(of: bar, upperBound: 300, plotWidth: 100)
-        XCTAssertEqual(drawn, 3, accuracy: 0.0001, "the floor here is worth three seconds")
-        XCTAssertEqual(WaterfallChartStyle.valueLabel(for: bar), DurationText.milliseconds(2))
-        XCTAssertFalse(WaterfallChartStyle.valueLabel(for: bar).contains("3"),
-                       "the label reports the measurement, never the drawn width")
-    }
-
     // MARK: - Page geometry
 
     /// The defect the owner found: rows that shrank to share the screen turned a scrollable
